requirements.txt
Flask==2.3.2
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4
gunicorn==21.2.0
scikit-surprise==1.1.1
